Reorganize modules into logical directories
