+++ src/js/app.js
@@ -1,21 +1,21 @@
 // STO Tools Keybind Manager - Main Application Controller
 // Coordinates all modules and handles global application state
-import store from './store.js'
-import eventBus from './eventBus.js'
-import STOPreferencesManager from './preferences.js'
-import STOAutoSyncManager from './autoSync.js'
-import { profileManagement } from './profileManagement.js'
-import { keyHandling } from './keyHandling.js'
-import { uiRendering } from './uiRendering.js'
-import { parameterCommands } from './parameterCommands.js'
-import { keyCapture } from './keyCapture.js'
-import { eventHandlers } from './eventHandlers.js'
-import { projectManagement } from './projectManagement.js'
-import { modeManagement } from './modeManagement.js'
-import { commandLibrary } from './commandLibrary.js'
-import { aliasView } from './aliasView.js'
-import { viewManagement } from './viewManagement.js'
-import { welcome } from './welcome.js'
+import store from './core/store.js'
+import eventBus from './core/eventBus.js'
+import STOPreferencesManager from './services/preferences.js'
+import STOAutoSyncManager from './services/autoSync.js'
+import { profileManagement } from './features/profileManagement.js'
+import { keyHandling } from './features/keyHandling.js'
+import { uiRendering } from './ui/uiRendering.js'
+import { parameterCommands } from './features/parameterCommands.js'
+import { keyCapture } from './ui/keyCapture.js'
+import { eventHandlers } from './ui/eventHandlers.js'
+import { projectManagement } from './services/projectManagement.js'
+import { modeManagement } from './ui/modeManagement.js'
+import { commandLibrary } from './features/commandLibrary.js'
+import { aliasView } from './ui/aliasView.js'
+import { viewManagement } from './ui/viewManagement.js'
+import { welcome } from './ui/welcome.js'
 
 export default class STOToolsKeybindManager {
   constructor() {
