+++ src/js/core/errors.js
