+++ src/js/features/vertigo_data.js
