+++ src/js/main.js
@@ -1,25 +1,25 @@
-import './constants.js'
-import eventBus from './eventBus.js'
+import './core/constants.js'
+import eventBus from './core/eventBus.js'
 import './data.js'
-import './errors.js'
+import './core/errors.js'
 import i18next from 'i18next'
 import en from '../i18n/en.json'
 import de from '../i18n/de.json'
 import fr from '../i18n/fr.json'
 import es from '../i18n/es.json'
-import STOStorage from './storage.js'
-import STOProfileManager from './profiles.js'
-import STOKeybindFileManager from './keybinds.js'
-import STOAliasManager from './aliases.js'
-import STOExportManager from './export.js'
-import STOModalManager from './modalManager.js'
-import STOUIManager from './ui.js'
-import STOCommandManager from './commands.js'
-import STOFileExplorer from './fileexplorer.js'
-import STOSyncManager from './sync.js'
-import VertigoManager, { VFX_EFFECTS } from './vertigo_data.js'
+import STOStorage from './services/storage.js'
+import STOProfileManager from './features/profiles.js'
+import STOKeybindFileManager from './features/keybinds.js'
+import STOAliasManager from './features/aliases.js'
+import STOExportManager from './features/export.js'
+import STOModalManager from './ui/modalManager.js'
+import STOUIManager from './ui/ui.js'
+import STOCommandManager from './features/commands.js'
+import STOFileExplorer from './ui/fileexplorer.js'
+import STOSyncManager from './services/sync.js'
+import VertigoManager, { VFX_EFFECTS } from './features/vertigo_data.js'
 import STOToolsKeybindManager from './app.js'
-import './version.js'
+import './ui/version.js'
 
 const stoStorage = new STOStorage()
 const settings = stoStorage.getSettings()
