+++ src/js/ui/fileexplorer.js
@@ -1,6 +1,6 @@
 // STO Tools Keybind Manager - File Explorer Modal
 // Provides a tree view of profiles/builds/aliases and preview of export files
-import eventBus from './eventBus.js'
+import eventBus from '../core/eventBus.js'
 
 export default class STOFileExplorer {
   constructor() {
