+++ src/js/ui/viewManagement.js
