+++ src/js/ui/welcome.js
