+++ tests/browser/user-workflows.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
-import eventBus from '../../src/js/eventBus.js'
+import eventBus from '../../src/js/core/eventBus.js'
 
 describe('Complete User Workflows', () => {
   let app, stoStorage, stoUI, stoExport, stoKeybinds
