+++ tests/integration/app-workflow.test.js
@@ -2,12 +2,12 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
 import fs from 'fs'
 import path from 'path'
 import '../../src/js/data.js'
-import eventBus from '../../src/js/eventBus.js'
-import STOStorage from '../../src/js/storage.js'
-import STOProfileManager from '../../src/js/profiles.js'
-import STOKeybindFileManager from '../../src/js/keybinds.js'
-import STOExportManager from '../../src/js/export.js'
-import STOUIManager from '../../src/js/ui.js'
+import eventBus from '../../src/js/core/eventBus.js'
+import STOStorage from '../../src/js/services/storage.js'
+import STOProfileManager from '../../src/js/features/profiles.js'
+import STOKeybindFileManager from '../../src/js/features/keybinds.js'
+import STOExportManager from '../../src/js/features/export.js'
+import STOUIManager from '../../src/js/ui/ui.js'
 import STOToolsKeybindManager from '../../src/js/app.js'
 
 describe('App Workflow Integration', () => {
