+++ tests/unit/aliases.test.js
@@ -5,10 +5,10 @@ import { join } from 'path'
 // Import real data first to ensure STO_DATA is available
 import '../../src/js/data.js'
 
-import eventBus from '../../src/js/eventBus.js'
+import eventBus from '../../src/js/core/eventBus.js'
 // Load the aliases module (it creates a global instance)
-import STOAliasManager from '../../src/js/aliases.js'
-import store, { resetStore } from '../../src/js/store.js'
+import STOAliasManager from '../../src/js/features/aliases.js'
+import store, { resetStore } from '../../src/js/core/store.js'
 
 // Load the real HTML
 const htmlContent = readFileSync(join(process.cwd(), 'src/index.html'), 'utf-8')
